ui_requirements.txt
PyQt5>=5.15
pyqtgraph>=0.13
numpy>=1.21
scipy>=1.7
pyserial>=3.5
